backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-dotenv
greenlet
